Testing web page redirection under GitHub
This is a simple test of both a 301 page redirection, and a META Refresh redirection.
If this is successful, then we will be able to redirect pages from where they
once were to where we will move them (such as the 1988 thru 2004 winners).

These HTML pages will go away after the test.