Fix markdown, html files and Makefile
In particular the lists were mangled; I made sure that the html files
have the IOCCC headers/footers by using rdiscount to another file and
then yanking/pasting the generated lists in the one file into the target
files.

I also removed terminals.markdown which was a mistake I made when I sent
a patch (so I assume) to the judges.

Makefile now refers to *.text instead of *.markdown and no longer refers
to judges file as that was only part of the preview repo for winners.

There might have been other changes I have made but everything *appears*
to be okay now.