Fix markdown and html files and Makefile
Much like the ferguson1/ fixes in the previous commit I have fixed the
lists in the markdown files and the html files have been updated too.

 I *believe* I got the chocolate-cake.html generated properly after
making a change of the chocolate-cake.text (because it's encrypted with
the Enigma and because the judges have their headers/footers I had to
take the generated file from ferguson1 - corrected - and then encrypt it
with the proper key).

Makefile now refers to *.text files rather than *.markdown.

There might be some other changes I forgot.