Change references of hint.text etc. to README.md
The following files were modified:

	modified:   2012/MANIFEST
	modified:   2012/hamano/README.md
	modified:   2014/MANIFEST
	modified:   2015/MANIFEST
	modified:   2018/MANIFEST
	modified:   2018/anderson/README.md
	modified:   2018/ciura/README.md
	modified:   2018/ferguson/FILES
	modified:   2018/ferguson/weasel.1
	modified:   2018/ferguson/weasel.man
	modified:   2018/ferguson/weasel.spec
	modified:   2019/MANIFEST
	modified:   2019/karns/README.md
	modified:   2020/MANIFEST
	modified:   2020/ferguson2/README.md
	modified:   2020/ferguson2/enigma.1
	modified:   2020/ferguson2/recode.1
	modified:   years-spoiler.html
	modified:   years.html

but I am not clear on some of the changes here. For instance some of the
text changed were html anchors and I'm not sure of the implications of
it other than the fact that changing the filenames (see commits
f4cfb073691e9447c09985d9308f112c57bc38bd and
4b65e324857472cef3bdbc6e5c2c33c8348d6df9) would break the links. Whether
more needs to be done or not I do not know.