Fix 1992/albert README
The README.md had an email to the judges from Leo (prior to being a
judge, presumably) which means that the index file would not end up with
information about the entry.

The old contents of the README.md were copied to albert.bug.md though
perhaps it should not be a markdown file. The name was arbitrarily
selected and a better name could certainly be contrived.