Fix 2020/ferguson[12] wrt years.html
Due to (at the time of publication of the 2020 winning entries) the
rename of '.markdown' to '.text' and earlier on not being sure of how to
get the correct footer in the files (except manually) some additional
files (.markdown) were added to these two entries. The .md files were
correct in that they have not only the correct footer but they also have
lists fixed (spaces between lines starting with '-').

As well in years.html some files referred to '.text' which no longer
exist either and others referred to '.markdown' files and still others
to '.text' files and those two extensions are wrong.

This commit deletes the following files:

	deleted:    2020/ferguson1/bugs.markdown
	deleted:    2020/ferguson1/gameplay.markdown
	deleted:    2020/ferguson1/spoilers.markdown
	deleted:    2020/ferguson1/terminals.markdown
	deleted:    2020/ferguson1/troubleshooting.markdown
	deleted:    2020/ferguson2/recode.markdown

which as I noted above were incorrect.

This commit is part of fixing years.html wrt missing files but since
they're all related to the same author in the same year (me in 2020) I
felt they belonged in a commit separate from the other fixes which are
more general.