Change references of 'EMail' to '[Ee]mail'
Depending on if it was the beginning of the sentence the EMail would
become 'Email' or 'email'.

Some references of 'Email' were changed to 'email' as well as those were
not in the beginning of a sentence.

A few places some consistency of capitalised email address (made lower
case) were changed to match other places of the same pattern.