Get 1984/lycklama to compile
To get it to compile we have to pass to the compiler -include unistd.h.

To get it to compile without warnings we had to add -Wno-parentheses to
the compiler invocation.

Updated the README to include an example invocation (using the program
source itself) as the old way no longer works (typing in long lines
directly).

Removed some comments that turn out to not be true (it does not need a
K&R compiler for instance).