Change 'OSX' to 'macOS' in 2020/kurdyukov2/README
This is because it is macOS and it's indeed what OS I explained how to
get the entry to compile under. It might not matter much but since we're
making things consistent (this same change might need to be done
elsewhere) and as correct as possible this commit seems good.