Minor fixes in 2020/burton
In the check_be.sh script the default expected prog_be to be in the path
but that's not likely so make it ./prog_be instead.

Updated README.md to change prog.be to be prog_be and also add a note
about it segfaulting without an argument at the judges' comments, close
to the top.

The Makefile also warns on the entry segfaulting on no argument. This
way the viewers can have a warning. I will look at doing this with other
entries that might segfault too but later on.

I also added some example invocations to highlight something Dave Burton
mentioned in his remarks. Of course if he's against this I have no
problem there but I think it enhances the README.