Fixes in top level README.md
Make links markdown format (since it's a markdown document).

Make some text bold and other text italic to help it all stand out.

Fix link to GitHub winner repo: it said it was the 'IOCCC winners GitHub
repository' but it actually linked to https://www.ioccc.org/index.html
which seems to be a mistake.