Attempt to fix conflicts
As some commits from yesterday and this morning caused conflicts with
commits I made last week I'm trying to fix it on the branch that I
believe introduced the changes. If this does not work I do not know what
I'll do but it have to be done another time.