Fixes in 1992/nathan/README.md
Restore absent-minded change of make all to make nathan.

Fix typos.

Fix formatting (the text in the sections is indented which makes it all
code blocks). This problem is probably because originally the file was
not README.md (though it was still considered markdown). Many other
README.md files have this problem which I will fix in time.