Get all 1986 entries to compile
1986/wall will not compile under the default gcc (as it's actually
clang) with macOS as it requires -traditional-cpp which unfortunately
clang does not support.

Also fixed many of the README.md files.

Removed 1986/bright/bright.orig.c (probably because it won't compile on
modern systems but maybe not; I did that change a while back but didn't
finish working on 1986 until just now).