Minor updates to 1984
Updated 1984/README.md with link to rules.txt.

Updated 1984/decot/README.md removing the note

    NOTE: This entry might not compile when using modern compilers.

as this was redundant (I already added more information in an earlier
commit). As well I moved the header Judges' comments below that note so
that the note about compiling is under the build section of the
README.md.