Update chocolate cake files (2020/ferguson[12])
There were some things I wanted to reiterate and stress even more that
are important notes about the cake.

Of course the ferguson2 files are still encrypted with a specific key
for the Enigma machine (one of the keys I provided in recode.md /
recode.html but which is a challenge though probably pointless since
it's also in ferguson1 unencrypted).

Updated 2020/ferguson2/.gitignore: added data files specific to the
encryption of the cake files so I can more easily fix it should I need
to update the files under 2020/ferguson2 (unlikely as that might be).