Fix in 1991/westley/ttt.sh
Don't assume '.' is in the path.

I note that at least under macOS this game does not work. I'll be
looking at the entries of this year at another time. I just need this
commit in to make some other changes unrelated.

The README.md was also updated - formatting fixes to turn it into a
markdown file proper.