Fix 1996/gandalf
As I love the theme and layout I kept the code that can no longer
compile in just commented out.

Fun fact: did you know that the original name of Gandalf was Bladorthin?
He was not the only character who originally had different names in The
Hobbit. In fact the original name of Thorin was Gandalf! (... I know the
former is true and I'm almost certain the latter is true but it's been a
long time since I've read it in The History of the Hobbit).