Add to / fix formatting in 1991/ant/README.md
In the judges' comments add some humour.

Add a try section.

Note that a file is needed for the program to do anything (that is a
file that exists) - not specifying an arg will exit without doing
anything at all.

Explain how to enter form-feed as this is important.

Fix formatting.