Add bugs.md back to repo
The rationale is that it'll become a html file on the main site and
having www.ioccc.org/BUGS.html seems less ideal than
www.ioccc.org/bugs.html though I suppose this is only subjectively
better esp as many distributions of software have a BUGS file (and in
fact I had such a file in 2020/ferguson1 but now it's lower case). Still
this can be debated later I suppose.