Greatly improve bugs.md
I have added sections as not all problems are the same. Some are not
supposed to be touched for instance and others have partial fixes. Some
will work with some compilers but not others. There is one entry (for
now) where things might appear to be bugs but are not (2020/ferguson1).

I added links to each entry along with the README.md file.

I have given tips to the problems in some cases.

Yes some of the entries refer to me and one entry even has a tip related
to the entry I worked on (2020/ferguson2) that helped me discover the
bug as using the Enigma machine can help identify problems (I gave
an example in the README.md file which I also linked to).