Fix 2001/cheong to work with clang
I had to make a new function (also four letter word to match the
formatting) that is allowed to have an int as the third arg to it. I
chose the word pain :-) as it's a pain that clang is this way without
providing any option to disable it.

Note that the arg must be an even number of digits. This threw me off
until I read the README.md because on different systems I got different
results for odd number of digits args despite the fact there should be
no difference here as the diff shows.