Actually fix 2001/ctk
It became obvious when I actually looked at the code (though I suspected
from output that it was using ANSI) that it was just no longer valid
ANSI escape sequences. Yusuke Endoh made a patch which I used (I have
used ANSI before but it's been a long while). Yusuke guessed it was a
linux thing (reasonable guess) but this also works for macOS. Thank you
Yusuke for your assistance!