Fix 2001/herrmann2 for clang
This was done by changing the types of the args in main() to be int and
char ** respectively and (where necessary) casting the argv (V) to int
which was the previous type.

I also noted that prior to and after the fix certain invocations
segfaulted though Yusuke Endoh did not have this problem (I linked to
his analysis of the entry).