Fix Westley's entries (dead link, .gitignore)
The main link was dead but looking at the Internet Wayback Machine I
found that he changed domains and it's still valid. One entry had a
different URL but with a quick check I could not determine that it had
anything to do with him so I changed it to his own domain (which I
verified is his).

The 2001/westley/.gitignore was also updated as it has other source
files that are generated by the commands in the try section.