1996 fixes and updates
Update the README.md files - at least with those not in a previous
commit. The updates mostly include formatting (make them markdown) but
probably also typos here and there. More probably has to be done with
these files.

The 1996/august/.gitignore was updated as well with a file that is
generated by one of the suggested commands.