Make 2000/jarijyrki compile without -I option
The includes for the X11 header files omitted the X11 and this means
running make will fail even if one has the files in a place that's in
the include path (default or via -I option).

Of course if one has X11 include files in a non-standard location (or
where the compiler doesn't look) they'll still have to use the -I
option.