2006/monge requires x86/x86_64 CPUs
The segfault I noted in 700f00eeade9720d1d4a6c1c1676e6b804cdc883 appears
to be because it requires x86/x86_64 architecture and I tested it on the
arm64 (MacBook Pro M1). I will still in time test it under linux to make
sure it works but that's for another time.

With this in mind the entry should at least theoretically work at least
if you have an x86/x86_64 CPU.