Update 2012/grothe/.gitignore
Several files that are generated in the try section were not excluded. I
note that some files that are there are in the repo and perhaps they
should not be - but I don't know for sure. They are binary files that
are created / modified by the entry but I'll hold off on removing those
and adding them to .gitignore for now at least.