Remove data files from 2012/grothe
These files are generated based on data from /dev/urandom and are
example uses. As anyone running this entry in a copy of the repo (fork
or otherwise) would get that the files are modified in git status (after
running the example commands) it seems to me that they should be deleted
and added to .gitignore which I have now done.

My guess (but it's just a guess) is that when the repo was set up these
commands were run and perhaps a git add . was done from the root
directory.