Fixes in 2015/burton
Format fixes (maybe typos?) in README.md.

Added calc.1 which is essentially what was calc.man but so one can
actually get expected formatting when using man ./calc.1. If I fixed any
typos they are not fixed in calc.man. I kept the commands the same so
one would have to change calc to ./prog.

Formatting and maybe typos in road.to.obscurity.