Fix 1988/phillipps for modern systems
The type of the second and third arg of main() were not char ** which
caused clang to have a fit. In gcc this is not a problem but it printed
gibberish nonetheless.

Changing the type of the second and third arg to compile with clang and
it printed the same (or similar?) gibberish as it did with gcc.

Adding a new function (pain() as it's annoying that clang requires
second, third and fourth args to main() to be 'char **'s even though in
this case it didn't work anyway) and calling it from main with the
appropriate args fixes the problem.