Add 1988/spinellis/spinellis.alt.c for clang
This was fun to do even though it's also a very simple one liner plus
special handling in the Makefile (the latter part is so you can use it
the exact same way as the original). The original code just #included
/dev/tty. This does something just as magical! It can be used like:

	make alt
	# some C code
	^D
	./spinellis

For instance:

	make alt
	#include <stdio.h>
	int main() { puts("hello world"); }
	^D
	./spinellis

Now those with clang can enjoy the entry without any functional
difference.

I have a very twisted point to make about the original version and this
version too: with a very slight change one can actually make a C++
version. Ick! :-)