Make 1992/gson not use gets() + update README.md
In the case of 1992/gson it was not as simple to change it to use
fgets() due to how the char * is set up. First it's set to argv (see
README.md for the more correct details) and then it's set to h which
was of size 256. Since the program requires the input be from the
command line I simply changed the buffer size to ARG_MAX+1 (which might
already be more than enough).

Fixed formatting in README.md.

Changed in Makefile the og to ag to match the description (and in the
places where og was used in the README.md I changed it to ag).

.gitignore also now has ag not og.

Possibly typos were fixed but I'm not sure now.