Change -ggdb3 to -O3 in 1992/gson/Makefile
This was accidentally committed. Before looking more closely at the code
I was hoping to find via the debugger where m was set up with a specific
size. This turned out to not be necessary after looking at the code but
nevertheless I enabled debug symbols and disabled the optimiser but this
was supposed to be temporary.