Delete faq.html
This was supposed to be done in commit
6213425e3cbe8eae011c8a576c98018e0a746852 but it failed because I had
local changes (due to testing the markdown) and instead of doing git rm
-f I did git restore etc. (or something like that .. don't remember now)
which seems to have messed it up. But with faq.md there we don't want
the html file as noted on GitHub (as I thought). Thus faq.html is now
gone.