Improvements and fixes to faq.md
Format fixes (besides line lengths other formatting changes were made in
hopes to make things clearer / easier to see / better).

Reword some answers a bit to hopefully make them better.

Typo fixes here and there. There is one I'm not sure if it's meant to be
a typo or rather if I should change it since it's actually a quote: that
from 1984/anonymous. Thus I have not changed it.